requirements.txt
torch>=1.7.1
tqdm
scikit-learn
pandas
plotly
dash
Pillow
pyyaml
av